+++ docs/source/notebooks/basic-demo.py
@@ -15,12 +15,13 @@
 # %% [markdown]
 # # Basic demo
 #
-# This notebook gives a basic demonstration of how to use Carpet - Concentrations.
+# This notebook gives a basic demonstration of how to use Carpet - concentrations.
+
+# %%
+# TODO: write demo
 
 # %%
 import carpet_concentrations
 
 # %%
-print(
-    f"You are using carpet_concentrations version {carpet_concentrations.__version__}"
-)
+print(f"{carpet_concentrations.__version__=}")